+++ backend/src/routes/activate-mail/activate.router.ts
@@ -1,5 +1,5 @@
 import { RoutPath } from '@common/enums/app/route-path.enum';
-import { activateMail } from '@controllers/mail-varification/verification';
+import { activateMail } from '@controllers/mail-verification/verification';
 import { Router } from 'express';
 
 const activateRouter = Router();
