+++ backend/src/routes/activation-link/activation-link.router.ts
@@ -1,5 +1,5 @@
 import { RoutPath } from '@common/enums/app/route-path.enum';
-import { sendVerificationLink } from '@controllers/mail-varification/send-link';
+import { sendVerificationLink } from '@controllers/mail-verification/send-link';
 import { Router } from 'express';
 
 const activateLinkRouter = Router();
