+++ backend/src/services/auth/auth.service.ts
@@ -3,8 +3,8 @@ import { prisma } from '@data/prisma-client';
 import { createToken } from '@helpers/helpers';
 import { bcryptHash, sendEmail } from '@helpers/helpers';
 import { User } from '@prisma/client';
-import { mailSend } from '@services/mail_verification/send.service';
-import { updateMailToken } from '@services/mail_verification/user-security.service';
+import { mailSend } from '@services/mail-verification/send.service';
+import { updateMailToken } from '@services/mail-verification/user-data.service/user-security';
 import bcrypt from 'bcrypt';
 import * as jwt from 'jsonwebtoken';
