+++ backend/src/services/mail-verification/send.service.ts
@@ -1,0 +1,8 @@
+const mailSend = (email: string): string => {
+  // Mock token, it must be genereted here
+  const token = 'rOiBUBf2pPvOHTNBeTR8CkLGokwHADhv';
+  console.log(`Sending email to ${email}`);
+  return token;
+};
+
+export { mailSend };
